Unified connect-button
